feat: add fe adapter
