+++ src/api/axios/cache.ts
@@ -1,14 +1,6 @@
 import { AxiosInstance, AxiosRequestConfig, AxiosResponse } from "axios";
 import { unstable_serialize } from "swr";
 
-declare module 'axios' {
-  interface AxiosRequestConfig<D = any> {
-    disableCache?: boolean;
-    wsApi?: 'unique' | true | undefined;
-    excludeMeta?: boolean;
-  }
-}
-
 interface Cache {
   get(key: string): Promise<any>;
