+++ src/api/axios/type.ts
@@ -1,0 +1,10 @@
+declare module 'axios' {
+  interface AxiosRequestConfig<D = any> {
+    disableCache?: boolean;
+    wsApi?: 'unique' | true | undefined;
+    excludeMeta?: boolean;
+    format?: 'compact';
+  }
+}
+
+export {};
